Added conditions for applying strategies
